@jkotas 2024-03-05
How does changing the struct to ref struct avoids the allocation?