@steveharter 2024-03-05
> How does changing the struct to ref struct avoids the allocation?

Yes - sorry. This was premature and got conflated with a previous commit that cached the ConstructorInfoExs in a static.